Test/Bug fixes, delay fix, store node number (#9)
Fix the current test suite since most of it was broken. This involved
mainly cleaning up functions to match the new protbufs, but also
included
- Fixing a bug when retrieving channels where the request hangs by
  rechecking after an empty response
- Fixing a similar bug with radio info
- Fixing a bug when setting latitude and longitude to take in int32
  instead of float

Also reworked the message process to store the node number when
initializing the radio struct. This probaby only helps tests, but could
be useful in some part of the CLI as well

This should fix issue #7